Change PageProps values to "any".
When working in the Laravel Jetstream template, using TypeScript, many errors are thrown due to PageProps values having type `unknown`. In projects where it is not possible to determine the types on the frontend, setting PageProps values to `any` allows for the benefits of TypeScript in the application, save for these props.

It is not currently possible to extend, and override the PageProps types in TypeScript, so I feel this will allow for easier adoption of TypeScript in Laravel Jetstream projects.