Update `useForm` types to accept interfaces
We still want to do some validation on the types so we use a conditional type in parameter position to validate that the data being passed through is correct. This isn’t the most ideal solution because the errors highlight the entire object instead of individual keys but it is an improvement over the status quo.