Restore command + click behavior on Inertia links in React (#2132)
In #1910 (d53b2b3), we switched to passing `event.nativeEvent` to
`shouldIntercept` in an attempt to consistently always pass native
browser events to this framework-agnostic helper. This actually
introduced a bug though, because `currentTarget` is _different_ on
React's synthetic event vs. the underlying native event. The correct
value for `currentTarget` lives on the synthetic event.

Since we are no longer relying on `event.which` (and this is the only
property we used to be concerned with that was not included on the
React synthetic event), we'll simply move back to passing the React
synthetic event to `shouldIntercept`.