Add optional chaining to prevent history issue.
We've hit a strange bug where this particular line causes an issue after using `Inertia::location()` - it tries to read the `page` parameter of `window.history.state` which i'm assuming is reset during an Inertia::location() call? 

Adding the optional chaining modifier (as already exists in another reference to this variable further up the page) would fix this for us.