do not pass url when storing scroll state to history
If there was a change to the url that interia was not aware of, such
as the modification of a query parameter, this would cause the current
url to be overwritten and reset to a previous value.

Since passing the url is optional, we can simply not pass it to avoid
overriding the url in any case.

See also https://github.com/inertiajs/inertia/issues/2233